+++ lib/src/inputs/src/pickers/month/range.dart
@@ -134,6 +134,7 @@ class ThemedMonthRangePicker extends StatefulWidget {
 }
 
 class _ThemedMonthRangePickerState extends State<ThemedMonthRangePicker> {
+  final TextEditingController _controller = TextEditingController();
   LayrzAppLocalizations? get i18n => LayrzAppLocalizations.maybeOf(context);
   late int _focusYear;
   bool get isDark => Theme.of(context).brightness == Brightness.dark;
@@ -165,12 +166,20 @@ class _ThemedMonthRangePickerState extends State<ThemedMonthRangePicker> {
   void initState() {
     super.initState();
     _overrideFocusMonth(isInit: true);
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      if (mounted) _controller.text = _parsedName ?? '';
+    });
   }
 
   @override
   void didUpdateWidget(ThemedMonthRangePicker oldWidget) {
     super.didUpdateWidget(oldWidget);
     _overrideFocusMonth();
+    if (widget.value != oldWidget.value) {
+      WidgetsBinding.instance.addPostFrameCallback((_) {
+        if (mounted) _controller.text = _parsedName ?? '';
+      });
+    }
   }
 
   void _overrideFocusMonth({bool isInit = false}) {
@@ -201,6 +210,7 @@ class _ThemedMonthRangePickerState extends State<ThemedMonthRangePicker> {
     }
 
     return ThemedTextInput(
+      controller: _controller,
       value: _parsedName ?? '',
       labelText: widget.labelText,
       label: widget.label,