+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 7.1.6
+
+- Bug fixes on multiple selects' widgets, now will display correctly the selected value.
+
+## 7.1.5
+
+- Added `richTextBuilder` to `ThemedColumn` to allow to build a `RichText` widget instead of a `Text` or `Widget`.
+
 ## 7.1.4
 
 - Added `enableBreadcumb` property on `ThemedLayout` to disable globally the breadcumb on the layout